HeapNode

Documentation

data Node a

The node type of the Binomial Heaps.

When doing QuickCheck homework, you should NOT access Node type -- the implementation that your properties check may not have Node!

Constructors

Node

Fields

value :: a

value n returns the element contained in node n.

The element can be of any type you specify: only key; key and value;...

rank :: Int

The rank of this node, used by binomial heap.

children :: [Node a]

The children of this node, used by binomial heap.